Move tower API into services dir
